move utility function
